% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button-package.R
\docType{package}
\name{button-package}
\alias{button}
\alias{button-package}
\title{button: Press a Button Now}
\description{
Text-based AI & Analytics Job Simulator.
}
\author{
\strong{Maintainer}: Charles Gray \email{first.last@example.com}

}
\keyword{internal}
