% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button_ui.R
\docType{data}
\name{button_ui}
\alias{button_ui}
\title{importFrom shiny fluidPage, textOutput
importFrom layout_sidebar, bs_theme, page_fluid}
\format{
An object of class \code{bslib_page} (inherits from \code{shiny.tag.list}, \code{list}) of length 5.
}
\usage{
button_ui
}
\description{
importFrom shiny fluidPage, textOutput
importFrom layout_sidebar, bs_theme, page_fluid
}
\keyword{datasets}
